Reorg, typo fixes
