+++ src/canvasUtils.js
@@ -1,6 +1,7 @@
 // @flow
 
 import memoize from 'memoize-one';
+import { INTERVAL_TIMES,MAX_INTERVAL_SIZE_PX } from './constants';
 
 // hidpi canvas: https://www.html5rocks.com/en/tutorials/canvas/hidpi/
 export function configureRetinaCanvas(
@@ -46,3 +47,38 @@ export function getCanvasMousePos(
 
   return { canvasMouseX, canvasMouseY };
 }
+
+// Time mark intervals vary based on the current zoom range and the time it represents.
+// In Chrome, these seem to range from 70-140 pixels wide.
+// Time wise, they represent intervals of e.g. 1s, 500ms, 200ms, 100ms, 50ms, 20ms.
+// Based on zoom, we should determine which amount to actually show.
+export function getTimeTickInterval(zoomLevel) {
+  let interval = INTERVAL_TIMES[0];
+  for (let i = 0; i < INTERVAL_TIMES.length; i++) {
+    const currentInterval = INTERVAL_TIMES[i];
+    const pixels = currentInterval * zoomLevel;
+    if (pixels <= MAX_INTERVAL_SIZE_PX) {
+      interval = currentInterval;
+    }
+  }
+  return interval;
+}
+
+export const cachedFlamegraphTextWidths = new Map();
+export const trimFlamegraphText = (context, text, width) => {
+  for (let i = text.length - 1; i >= 0; i--) {
+    const trimmedText = i === text.length - 1 ? text : text.substr(0, i) + '…';
+
+    let measuredWidth = cachedFlamegraphTextWidths.get(trimmedText);
+    if (measuredWidth == null) {
+      measuredWidth = context.measureText(trimmedText).width;
+      cachedFlamegraphTextWidths.set(trimmedText, measuredWidth);
+    }
+
+    if (measuredWidth <= width) {
+      return trimmedText;
+    }
+  }
+
+  return null;
+};
